Split students into separate component
