+++ client/src/components/classesPage/components/Students.js
@@ -1,0 +1,101 @@
+import React, { useState, useEffect } from 'react';
+import { withStyles } from '@material-ui/core/styles';
+import {
+  Grid,
+  Typography,
+  Card,
+  Checkbox,
+  Button,
+  FormGroup,
+  Fab
+} from '@material-ui/core';
+import { GroupAdd } from '@material-ui/icons';
+
+const styles = theme => ({
+  studentList: {
+    display: 'flex',
+    flexFlow: 'column wrap',
+    border: `1px solid ${theme.palette.secondary.main}`,
+    flexWrap: 'wrap',
+    [theme.breakpoints.only('xs')]: {
+      width: '70%'
+    },
+    maxHeight: theme.spacing.unit * 50,
+    padding: theme.spacing.unit * 2
+  }
+});
+
+function students(props) {
+  const { classes } = props;
+  const [addedStudents, setAddedStudents] = useState([]);
+
+  function selectStudent(e) {
+    const studentId = e.target.value;
+    let updatedStudents = props.selectedStudents;
+    if (e.target.checked) {
+      updatedStudents = props.selectedStudents.concat(studentId);
+    } else {
+      updatedStudents = props.selectedStudents.filter(s => s !== studentId);
+    }
+    props.setSelectedStudents(updatedStudents);
+  }
+
+  return (
+    <>
+      <Typography variant="h6" className={classes.title} gutterBottom>
+        Current Students
+      </Typography>
+      <Card className={classes.studentList}>
+        {props.students.map((s, i) => (
+          <Grid key={i}>
+            <span>{`${s.name}`}</span>
+            <Checkbox
+              color="secondary"
+              value={`${s.student_id}`}
+              checked={props.selectedStudents.includes(s.student_id)}
+              onClick={e => selectStudent(e)}
+            />
+          </Grid>
+        ))}
+        {addedStudents.map((s, i) => (
+          <Grid key={i}>
+            <span style={{ fontWeight: 'bold' }}>{`${s.first_name} ${
+              s.last_name
+            }`}</span>
+            <Checkbox
+              color="primary"
+              value={`${s.student_id}`}
+              checked={props.selectedStudents.includes(s.student_id)}
+              onClick={e => selectStudent(e)}
+            />
+          </Grid>
+        ))}
+      </Card>
+      <FormGroup className={classes.settingsBox}>
+        <Typography variant="body1" gutterBottom>
+          Add a Student
+        </Typography>
+        {props.makeInput('email', 'Email')}
+        {props.makeInput('first_name', 'First Name')}
+        {props.makeInput('last_name', 'Last Name')}
+        <Fab
+          elevation={20}
+          aria-label="Add"
+          className={classes.btn}
+          onClick={e => props.addStudent(e)}
+        >
+          <GroupAdd />
+        </Fab>
+      </FormGroup>
+      <Grid className={classes.buttonBox}>
+        {props.selectedStudents.length ? (
+          <Button variant="outlined" onClick={props.dropStudents}>
+            Remove selected from class
+          </Button>
+        ) : null}
+      </Grid>
+    </>
+  );
+}
+
+export default withStyles(styles, { withTheme: true })(students);
